browser: start shared activation method
We will share code in the activate() functions
between the node and browser implementations.

This starts the transition process and we will progressively
refactor the code in the following commits.
